Fix broken login
The bug was introduced in https://github.com/eXist-db/dashboard/commit/af4cc34b8622a75b5c35289f32e4fab62ccb0207 - a missing curly brace.

This same bug is also fixed in #46, but since that PR is not yet pulled in — there are additional changes in that PR that appear to require deeper review — can we just fix this one now?  It's a no-brainer.