Circumvention of problem in public-repo
There is a pull request for a fix under review, but adding this one line of code makes the client compatible with the active public-repo code today.

Due to a probleem in the Content-Type of data from the vanilla public-repo, the return type is text/plain instead of application/xml. When this happens, the oneliner kicks in and does parsing into xml so the rest code works as expected.